+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 1.x branch
 ## 1.3 branch
 ### 1.3.5
+* added all classes for event management and the `EventAssertTrait` to assert
+    whether events were fired or not;
 * added `string_contains()` global function;
 * fixed bug for `rtr()` global function;
 * fixed bug for `objects_map()` global function. It now works if the class